+++ huskarl-core/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.1] - 2026-05-21
+
+### Additions
+
+ - Adds BoxedAeadCipher - boxed value that implements both encryption and decryption.
+
 ## [0.5.0] - 2026-05-06
 
 ### Additions